Move AdminLessonInputs to its own folder
