+++ components/admin/lessons/AdminLessonInputs/adminLessonInputs.module.scss
@@ -1,4 +1,4 @@
-@use '_variables.module';
+@use '../../../../scss/_variables.module';
 
 @mixin message($bg, $color) {
   display: flex;
